remove_files is free from russh
